Rename api package to client.
This moves all external API calls and logic from the api package
to the client package. This name makes more sense and is more
consistant with other go packages and allows us to in the future
create an API for replicator.

Closes #42
